+++ cypress/e2e/toast_injector.cy.ts
@@ -4,7 +4,7 @@ import { HOT_TOAST_DEFAULT_TIMEOUTS } from '../../projects/ngxpert/hot-toast/src
 describe('Test hot toasts - component', () => {
   it('should show and hide component toast', () => {
     cy.get('#injector').click();
-    cy.get('hot-toast').as('componentToast');
+    cy.get('hot-toast-component').as('componentToast');
 
     cy.get('@componentToast').should('contain', 'I love Angular 🔥 Hot Toasts!');
     cy.wait(HOT_TOAST_DEFAULT_TIMEOUTS.blank);