+++ cypress/e2e/toast_snackbar.cy.ts
@@ -5,7 +5,7 @@ import { HOT_TOAST_DEFAULT_TIMEOUTS } from '../../projects/ngxpert/hot-toast/src
 describe('Test hot toasts - snackbar', () => {
   it('should show and toast with snackbar theme', () => {
     cy.get('#snackbar').click();
-    cy.get('hot-toast').as('snackbarToast');
+    cy.get('hot-toast-component').as('snackbarToast');
 
     cy.get('@snackbarToast').should('contain', 'Snackbar');
     cy.get('@snackbarToast').find('.hot-toast-bar-base-container').should('have.css', 'bottom', '0px');