+++ cypress/e2e/toast_warning.cy.ts
@@ -5,7 +5,7 @@ import { HOT_TOAST_DEFAULT_TIMEOUTS } from '../../projects/ngxpert/hot-toast/src
 describe('Test hot toasts - success', () => {
   it('should show and hide success toast', () => {
     cy.get('#warning').click();
-    cy.get('hot-toast').as('warningToast');
+    cy.get('hot-toast-component').as('warningToast');
 
     cy.get('@warningToast').should('contain', 'Please be cautious!');
     cy.wait(HOT_TOAST_DEFAULT_TIMEOUTS.warning);