+++ projects/ngxpert/hot-toast/src/public-api.ts
@@ -4,6 +4,8 @@
 
 export * from './lib/hot-toast.service';
 export * from './lib/hot-toast.provide';
+export * from './lib/hot-toast-http-interceptor';
+export * from './lib/hot-toast-http-interceptor.model';
 export * from './lib/hot-toast.model';
 export * from './lib/hot-toast-ref';
 export * from './lib/hot-toast-builder';